Основной код игры разбит на файлы
