+++ data.py
@@ -1,0 +1,26 @@
+towers_data = [
+    {
+        'name': 'Cannon Tower',
+        'icon': 'assets/tower1.png',
+        'rate_of_fire': 2,
+        'damage': 10,
+        'visibility_zone': 200,
+        'cost': 200  # Стоимость
+    },
+    {
+        'name': 'Archer Tower',
+        'icon': 'assets/tower2.png',
+        'rate_of_fire': 5,
+        'damage': 2,
+        'visibility_zone': 300,
+        'cost': 150
+    },
+    {
+        'name': 'Wtf is this tower',
+        'icon': 'assets/tower1.png',
+        'rate_of_fire': 25,
+        'damage': 0.2,
+        'visibility_zone': 500,
+        'cost': 300
+    }
+]
